# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/vehicles": {
      "get": {
        "operationId": "VehiclesController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vehicles"
        ]
      },
      "post": {
        "operationId": "VehiclesController_createMine",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Vehicles"
        ]
      }
    },
...
    "/api/vehicles/{id}": {
      "patch": {
        "operationId": "VehiclesController_updateMine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vehicles"
        ]
      },
      "delete": {
        "operationId": "VehiclesController_removeMine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vehicles"
        ]
      }
    },
...
    "/api/v1/admin/vehicles": {
      "get": {
        "operationId": "AdminV1Controller_listVehicles",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search by make/model/plate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminV1"
        ]
      }
    },
...
    "/api/geo/trips/{tripId}/location": {
      "patch": {
        "operationId": "GeoController_updateTripDriverLocation",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Geo"
        ]
      },
      "get": {
        "operationId": "GeoController_getTripDriverLocation",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Geo"
        ]
      }
    },
    "/api/geo/trip/{tripId}/eta": {
      "get": {
        "operationId": "GeoController_getTripEta",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Geo"
        ]
      }
    },
...
  },
  "info": {
    "title": "InterCity API",
    "description": "Backend API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateVehicleDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateVehicleDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateDriverLocationDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}